software/master_node/requirements.txt
opencv-python>=4.8.0
pyzbar>=0.1.9
numpy>=1.24.0
pyserial>=3.5
pygame>=2.5.0
pyyaml>=6.0
